Replace float with double

+++ src/libcell/Reaction.cpp
@@ -50,7 +50,7 @@ Reaction::Type inferType(const std::optional<DiscType>& educt2, const std::optio
 }
 
 Reaction::Reaction(const DiscType& educt1, const std::optional<DiscType>& educt2, const DiscType& product1,
-                   const std::optional<DiscType>& product2, float probability)
+                   const std::optional<DiscType>& product2, double probability)
     : educt1_(educt1)
     , educt2_(educt2)
     , product1_(product1)
@@ -122,12 +122,12 @@ void Reaction::setProduct2(const DiscType& product2)
     product2_ = product2;
 }
 
-float Reaction::getProbability() const
+double Reaction::getProbability() const
 {
     return probability_;
 }
 
-void Reaction::setProbability(float probability)
+void Reaction::setProbability(double probability)
 {
     if (probability < 0 || probability > 1)
         throw ExceptionWithLocation("Probability must be between 0 and 1");
@@ -142,8 +142,8 @@ const Reaction::Type& Reaction::getType() const
 
 void Reaction::validate() const
 {
-    const float eductMass = educt1_.getMass() + (educt2_.has_value() ? educt2_->getMass() : 0);
-    const float productMass = product1_.getMass() + (product2_.has_value() ? product2_->getMass() : 0);
+    const double eductMass = educt1_.getMass() + (educt2_.has_value() ? educt2_->getMass() : 0);
+    const double productMass = product1_.getMass() + (product2_.has_value() ? product2_->getMass() : 0);
 
     if (eductMass != productMass)
         throw ExceptionWithLocation(toString(*this) + ": Product- and educt masses need to be identical");
